+++ my-app/src/board/board.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "../styles/variables.scss";
 
 .board-row:after {
     clear: both;
