+++ my-app/src/index.scss
@@ -1,25 +1,5 @@
-@import url(./board/board.scss);
-@import url(./square/square.scss);
+// available to import only generic files
 
-body {
-  font: 14px "Century Gothic", Futura, sans-serif;
-  margin: 20px;
-}
-
-ol,
-ul {
-  padding-left: 30px;
-}
-
-.status {
-  margin-bottom: 10px;
-}
-
-.game {
-  display: flex;
-  flex-direction: row;
-}
-
-.game-info {
-  margin-left: 20px;
-}
+@import './styles/common.scss';
+@import './styles/mixins.scss';
+@import './styles/variables.scss'
