+++ my-app/src/square/square.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './square.scss'
 
 function Square(props) {
     return (
