+++ my-app/src/styles/common.scss
@@ -1,0 +1,24 @@
+
+body {
+    font: 14px "Century Gothic", Futura, sans-serif;
+    margin: 20px;
+  }
+  
+  ol,
+  ul {
+    padding-left: 30px;
+  }
+  
+  .status {
+    margin-bottom: 10px;
+  }
+  
+  .game {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .game-info {
+    margin-left: 20px;
+  }
+
