+++ my-app/src/styles/variables.scss
@@ -1,0 +1,5 @@
+
+$color-background: #fbfbfb;
+$border-color-square: #999;
+$size-square: 34px;
+$font-square:24px;
