--- my-app/src/variables.scss
@@ -1,5 +1,0 @@
-
-$background: #fbfbfb;
-$border: #999;
-$square-size: 34px;
-$square-font:24px;
